backend/requirements.txt
asgiref==3.4.1
autopep8==1.5.7
Django==3.0.3
django-cors-headers==3.7.0
djangorestframework==3.12.4
Pillow==7.0.0
pycodestyle==2.7.0
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
typing-extensions==3.10.0.0
